refactor(core): split model types from runtime helpers
